+++ database/migrations/2022_09_01_000000_add_company_id_in_all_table.php
@@ -1,13 +1,11 @@
 <?php
 
 use App\Models\GlobalSetting;
-use App\Observers\CompanyObserver;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 use App\Models\Company;
-use Illuminate\Support\Str;
 
 return new class extends Migration {
 
@@ -27,7 +25,7 @@ return new class extends Migration {
         $first = Company::pluck('id')->first();
         // Delete if more than one companies in database for worksuite
         // Delete all other companies except first
-        if ($count > 1 && $first === 1) {
+        if ($count > 1 && $first === 1 && isWorksuite()) {
             Company::where('id', '<>', $first)->delete();
         }
 
@@ -110,7 +108,7 @@ return new class extends Migration {
                 $table->unique(['name', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
 
         try {
@@ -119,7 +117,7 @@ return new class extends Migration {
                 $table->unique(['name', 'module_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
 
         try {
@@ -131,7 +129,7 @@ return new class extends Migration {
                 $table->unique(['slack_username', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('estimates', function (Blueprint $table) {
@@ -139,7 +137,7 @@ return new class extends Migration {
                 $table->unique(['estimate_number', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('invoices', function (Blueprint $table) {
@@ -147,7 +145,7 @@ return new class extends Migration {
                 $table->unique(['invoice_number', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('lead_sources', function (Blueprint $table) {
@@ -155,7 +153,7 @@ return new class extends Migration {
                 $table->unique(['type', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('lead_status', function (Blueprint $table) {
@@ -163,7 +161,7 @@ return new class extends Migration {
                 $table->unique(['type', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('payments', function (Blueprint $table) {
@@ -174,7 +172,7 @@ return new class extends Migration {
                 $table->unique(['transaction_id', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('push_subscriptions', function (Blueprint $table) {
@@ -182,7 +180,7 @@ return new class extends Migration {
                 $table->unique(['endpoint', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('taskboard_columns', function (Blueprint $table) {
@@ -190,7 +188,7 @@ return new class extends Migration {
                 $table->unique(['column_name', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('ticket_channels', function (Blueprint $table) {
@@ -198,7 +196,7 @@ return new class extends Migration {
                 $table->unique(['channel_name', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('ticket_types', function (Blueprint $table) {
@@ -206,7 +204,7 @@ return new class extends Migration {
                 $table->unique(['type', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
         try {
             Schema::table('users', function (Blueprint $table) {
@@ -214,7 +212,7 @@ return new class extends Migration {
                 $table->unique(['email', 'company_id']);
             });
         } catch (\Exception $e) {
-            echo $e->getMessage();
+
         }
     }
